requirements.txt
pytest==9.0.2
pytest-xdist==3.5.0
pytest-html==4.2.0
pytest-cov==7.0.0
pytest-testmon>=2.0.0
pytest-asyncio>=1.3.0
pywinauto==0.6.8
PyYAML==6.0.2
allure-pytest==2.13.5
psutil==6.1.1
python-dotenv==1.0.1
pillow==11.1.0
